task0

+++ doubly_linked_lists/0-print_dlistint.c
@@ -10,11 +10,11 @@
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-	int i = 0;
+	size_t i = 0;
 
 	if (h == NULL)
 		return (0);
-	while (h && h ->prev)
+	while (h && h->prev)
 		h = h->prev;
 
 	while (h != NULL)
